% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{calc_risk}
\alias{calc_risk}
\title{Calculate Risk}
\usage{
calc_risk(risk, lambda, meanlog, sdlog, runs = 1e+05)
}
\arguments{
\item{risk}{Risk name (a string)}

\item{lambda}{Parameter passed to \code{rpois()}}

\item{meanlog}{Parameter passed to \code{rlnorm()}}

\item{sdlog}{Parameter passed to \code{rlnorm()}}

\item{runs}{Number of simulations}
}
\value{
A tibble of \code{runs} rows containing the number of events (\code{events}), and the total value
of losses (\code{losses}) for each run.
}
\description{
Calculate risk using a Poisson log-normal model. The number of events per time period is
calculated using \code{rpois(runs, lambda)} and total loss for each time period using
\code{sum(rlnorm(events, meanlog, sdlog))}.
}
