% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{lnorm_param}
\alias{lnorm_param}
\title{Calculate Log Normal Parameters}
\usage{
lnorm_param(p05, p95, p50)
}
\arguments{
\item{p05}{the "low" value, the 0.05 quantile.}

\item{p95}{the "high" value, the 0.95 quantile.}

\item{p50}{the most likely value, the 0.5 quantile.}
}
\value{
a list including the \code{meanlog} and \code{sdlog} parameters, as well as the fractional
difference between the estimated and actual median (0.5 quantile), \code{mdiff}.
}
\description{
Calculate the parameters of a log normal distribution from estimated 0.05 and 0.95 quantile. Also
calculate the fractional difference between the estimated median value (0.5 quantile) and the
median value calculated from the 0.05 and 0.95 values.
}
\examples{
lnorm_param(100000, 20000000, 1425000)

}
