% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{security-package}
\alias{security}
\alias{security-package}
\title{security: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jabenninghoff.github.io/security/}
  \item \url{https://github.com/jabenninghoff/security}
  \item Report bugs at \url{https://github.com/jabenninghoff/security/issues}
}

}
\author{
\strong{Maintainer}: John Benninghoff \email{jbenninghoff@mac.com} (\href{https://orcid.org/0000-0002-6230-4742}{ORCID})

}
\keyword{internal}
